Update: use of gfServer to perform the analysis
gfServer (part of the blatSuite) speeds up the alignment of the
sequences to the reference. Here an instance of the server is generated
and the fasta files are rapidily aligned.
The Mitochondrial filter now checks for similarity to chrM, similarly
to what the Ribosomal filter does.
TO DO: update the documentation and the Requirements for FusionSeq (and
.fusionseqrc)